pyproject.toml
[build-system]
requires = ["setuptools>=40.6.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "commit-tracker"
version = "0.1.0"
description = "Commit Tracker analyzes Git repositories to present comprehensive statistics."
authors = [{name = "André F. Costa", email = "afmcosta@gmail.com"}]
license = {file = "LICENSE.txt"}
readme = "README.md"
dependencies = [
    "dash",
    "GitPython",
    "pandas",
    "plotly",
    "dash-bootstrap-components"
]

[tool.setuptools.scripts]
commit-tracker = "commit_tracker.__main__:main"

[project.entry-points."commit_tracker.plugins"]
clone_remote_repo = "commit_tracker.git_stats:clone_remote_repo"
extract_git_stats = "commit_tracker.git_stats:extract_git_stats"
extract_branches_info_new = "commit_tracker.git_stats:extract_branches_info_new"
categorize_commit_type = "commit_tracker.git_stats:categorize_commit_type"
